[refactor] : 리팩토링 및 라우팅 구현
1. 이슈 상세 페이지 라우팅이 가능하도록 하였습니다.
2. IssueDeatilPage => IssueDetailPage로 파일 이름 변경하였습니다.
