+++ client/src/components/issue/detail/TitleDiv.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import { IssueContext } from '../../../pages/IssueDeatilPage';
+import { IssueContext } from '../../../pages/IssueDetailPage';
 import GreyButton from '../../common/GreyButton';
 
 const Div = styled.div`
